+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add scope selection after search
 * Fix keyboard scroll on token list
 * Fix focus issues
+* Add onScopesChange method
 
 ## 0.8.3 - 2019-01-10